modified Payout and Remittance Charge Commission in new format.

+++ setting-service/src/main/java/com/toshaniFintech/user_service/dto/request/RemittanceChargeCommissionRequest.java
@@ -5,23 +5,21 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 public class RemittanceChargeCommissionRequest {
     @NotNull(message = "Min amount is required")
     @DecimalMin(value = "0.0", inclusive = true, message = "Min amount must be >= 0")
-    private BigDecimal minAmount;
+    private Double minAmount;
 
     @NotNull(message = "Max amount is required")
     @DecimalMin(value = "0.0", inclusive = true, message = "Max amount must be >= 0")
-    private BigDecimal maxAmount;
+    private Double maxAmount;
 
     @NotBlank(message = "Charge type is required")
     private String chargeType;
 
     @NotNull(message = "Charge is required")
     @DecimalMin(value = "0.0", inclusive = true, message = "Charge must be >= 0")
-    private BigDecimal charge;
+    private Double charge;
 
 }
