+++ setting-service/src/main/java/com/toshaniFintech/user_service/model/RemittanceChargeCommissionModel.java
@@ -7,8 +7,8 @@ import java.math.BigDecimal;
 @Data
 public class RemittanceChargeCommissionModel {
     private String slNo;
-    private BigDecimal minAmount;
-    private BigDecimal maxAmount;
+    private Double minAmount;
+    private Double maxAmount;
     private String chargeType;
-    private BigDecimal charge;
+    private Double charge;
 }
